refactor(server): extract resolveSession + wire account claim in V1ClaimController
