+++ apps/server/src/interface/PseudoController.ts
@@ -5,23 +5,9 @@ import { PrismaUserRepository } from "../infrastructure/PrismaUserRepository";
 import { PrismaScoreRepository } from "../infrastructure/PrismaScoreRepository";
 import { checkRateLimit } from "../infrastructure/InMemoryRateLimiter";
 import { validatePseudo } from "../domain/Pseudo";
+import { resolveSession } from "./resolveSession";
 
 const FORMAT_ERROR = "3 à 20 caractères — lettres, chiffres, underscore uniquement";
-
-async function resolveSession(req: Request): Promise<string | null> {
-  const authUrl = process.env.AUTH_URL ?? "http://localhost:8888";
-  try {
-    const headers: Record<string, string> = {};
-    if (req.headers.cookie) headers["cookie"] = req.headers.cookie;
-    if (req.headers.authorization) headers["authorization"] = req.headers.authorization;
-    const res = await fetch(`${authUrl}/api/auth/get-session`, { headers });
-    if (!res.ok) return null;
-    const data = (await res.json()) as { user?: { id?: string } } | null;
-    return data?.user?.id ?? null;
-  } catch {
-    return null;
-  }
-}
 
 export class PseudoController {
   static async set(req: Request, res: Response): Promise<void> {
